requirements.txt
discord.py>=2.3.2
psycopg[binary]>=3.1
python-dotenv>=1.0.1
